chore: Clean up and simplify URL using query-string
