+++ components/DetailPathBar/DetailPathBar.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRouter } from 'next/router';
 import Image from 'next/image';
-import icon from '../public/arrow.svg';
+import icon from '../../public/arrow.svg';
 import { useStyles } from './styles';
 
 const DetailPathBar: React.FC = () => {
