+++ utils/common.js
@@ -1,0 +1,1 @@
+export const HOME_PAGE = 'http://localhost:3000/'
